Use our own cookie for storing relay_state, etc.
With browsers defaulting to SameSite: Lax for cookies, simply storing
the relay_state in the Session will fail when the IdP does a POST back
to our original site (i.e. Service Provider). Cookies are *not* included
when SameSite is set to lax, and so the relay_state will never match.

Instead, we set our own cookie with SameSite set to None.

The cookie is encrypted and given a max age of 5 minutes (in case they
need to go through their own auth flow before being redirect).

Since SameSite is None, Secure must be set to true; JS does not need
this cookie (and couldn't read it anyway), so Http Only is also set to
true.